# docs/contracts/openapi/deployer.openapi.json
{
  "components": {
    "schemas": {
...
      "Body_validateHierarchy": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "description": "Hierarchy JSON file",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_validateHierarchy",
        "type": "object"
      },
...
      "Body_validateSceneConfig": {
        "properties": {
          "hierarchy_file": {
            "contentMediaType": "application/octet-stream",
            "description": "Hierarchy JSON for cross-reference (optional)",
            "title": "Hierarchy File",
            "type": "string"
          },
          "scene_file": {
            "contentMediaType": "application/octet-stream",
            "description": "Scene config file (scene.json or 3DScenesConfiguration.json)",
            "title": "Scene File",
            "type": "string"
          }
        },
        "required": [
          "scene_file"
        ],
        "title": "Body_validateSceneConfig",
        "type": "object"
      },
...
      "Body_validateUserConfig": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "description": "config_user.json file",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_validateUserConfig",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HttpErrorEnvelope": {
        "description": "Actual FastAPI error shape returned by HTTPException handlers.",
        "example": {
          "detail": "The requested operation could not be completed."
        },
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "items": {},
                "type": "array"
              }
            ],
            "description": "Human-readable error text or a structured operation-specific error. Clients must preserve structured detail objects when present.",
            "title": "Detail"
          }
        },
        "required": [
          "detail"
        ],
        "title": "HttpErrorEnvelope",
        "type": "object"
      },
...
      "ProviderEnum": {
        "enum": [
          "aws",
          "azure",
          "gcp",
          "google"
        ],
        "title": "ProviderEnum",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Internal API for deploying, destroying, and inspecting Digital Twin resources.",
    "title": "Digital Twin Manager API",
    "version": "2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/validate/hierarchy": {
      "post": {
        "description": "**Purpose:** Validates hierarchy JSON for L4 Digital Twins (AWS IoT TwinMaker or Azure ADT).\n\n**When to call:** To validate aws_hierarchy.json or azure_hierarchy.json before deployment.\n\n**AWS format:** Array of entity definitions with type, id, and optional children.\n**Azure format:** Object with header, models, twins, and relationships arrays.",
        "operationId": "validateHierarchy",
        "parameters": [
          {
            "description": "L4 provider (aws or azure)",
            "in": "query",
            "name": "provider",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ProviderEnum",
              "description": "L4 provider (aws or azure)"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_validateHierarchy"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Hierarchy is valid"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Invalid request. Correct the supplied parameters or artifact and retry."
                },
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorEnvelope"
                }
              }
            },
            "description": "Bad Request - The request is invalid for this operation"
          },
          "413": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "The uploaded artifact exceeds the configured size limit."
                },
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorEnvelope"
                }
              }
            },
            "description": "Payload Too Large - The uploaded artifact exceeds the endpoint limit"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Internal server error. Check logs."
                },
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error - An unexpected error occurred"
          }
        },
        "summary": "Validate hierarchy JSON for L4 provider",
        "tags": [
          "Validation"
        ]
      }
    },
...
    "/validate/scene-config": {
      "post": {
        "description": "**Purpose:** Validates scene configuration for L4 3D visualization.\n\n**When to call:** To validate scene.json (AWS) or 3DScenesConfiguration.json (Azure).\n\n**Azure:** Validates JSON schema and cross-references primaryTwinID against hierarchy twins.",
        "operationId": "validateSceneConfig",
        "parameters": [
          {
            "description": "L4 provider (aws or azure)",
            "in": "query",
            "name": "provider",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ProviderEnum",
              "description": "L4 provider (aws or azure)"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_validateSceneConfig"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Scene config is valid"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Invalid request. Correct the supplied parameters or artifact and retry."
                },
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorEnvelope"
                }
              }
            },
            "description": "Bad Request - The request is invalid for this operation"
          },
          "413": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "The uploaded artifact exceeds the configured size limit."
                },
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorEnvelope"
                }
              }
            },
            "description": "Payload Too Large - The uploaded artifact exceeds the endpoint limit"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Internal server error. Check logs."
                },
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error - An unexpected error occurred"
          }
        },
        "summary": "Validate scene configuration with hierarchy cross-reference",
        "tags": [
          "Validation"
        ]
      }
    },
...
    "/validate/user-config": {
      "post": {
        "description": "**Purpose:** Validates config_user.json for L4/L5 browser access.\n\n**When to call:** To validate the platform identity before deployment.\n\n**Phase 8:** Trusted profile and L4/L5 context enables cross-cloud identity, Entra principal, AWS intent, and GCP CIDR checks.\n\n**Historical Azure:** Email must use a verified *.onmicrosoft.com domain.",
        "operationId": "validateUserConfig",
        "parameters": [
          {
            "description": "Historical provider or selected L5 provider",
            "in": "query",
            "name": "provider",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ProviderEnum",
              "description": "Historical provider or selected L5 provider"
            }
          },
          {
            "in": "query",
            "name": "architecture_profile_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Architecture Profile Id"
            }
          },
          {
            "in": "query",
            "name": "architecture_profile_version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Architecture Profile Version"
            }
          },
          {
            "in": "query",
            "name": "layer_4_provider",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Layer 4 Provider"
            }
          },
          {
            "in": "query",
            "name": "layer_5_provider",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Layer 5 Provider"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_validateUserConfig"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "User config is valid"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Invalid request. Correct the supplied parameters or artifact and retry."
                },
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorEnvelope"
                }
              }
            },
            "description": "Bad Request - The request is invalid for this operation"
          },
          "413": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "The uploaded artifact exceeds the configured size limit."
                },
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorEnvelope"
                }
              }
            },
            "description": "Payload Too Large - The uploaded artifact exceeds the endpoint limit"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Internal server error. Check logs."
                },
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error - An unexpected error occurred"
          }
        },
        "summary": "Validate config_user.json for platform user",
        "tags": [
          "Validation"
        ]
      }
    },
...
  },
...
}